Simulate air conditioner by playing a sound
This sound is played for now on aplay (to not add a new dep). However, we'll
use a module as soon as we can switch to python3. Play it in a separate thread
to be able to pkill the sound when reentering the event.
The sound play continously until killed.